pyproject.toml
[tool.poetry]
name = "lila-game-bot"
version = "0.1.0"
description = ""
authors = ["Nikolay Adamov <nikolay.adamov@servolux.by>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
python-dotenv = "^1.2.1"
aiosqlite = "^0.22.1"
aiogram = "^3.25.0"
pydantic = "^2.12.5"
pydantic-settings = "^2.13.1"
aiocron = "^2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
